Rename mod Add configuration screen and sets of translations.
