+++ README_EN.MD
@@ -1,4 +1,4 @@
-# Terra-EntityTracking HUD (English|[简体中文](README.MD))
+# Entity Tracker (English|[简体中文](README.MD))
 This is a simple HUD for tracking entities in Minecraft.
 ### How it works
 Using Mercator projection, it displays entities' positions in 2D, surrounding the crosshair.
