+++ gradle.properties
@@ -13,11 +13,11 @@ neo_version=21.1.116
 neo_version_range=[21.1.0,)
 loader_version_range=[4,)
 
-mod_id=terra_ethud
-mod_name=Terra-EntityTracking HUD
+mod_id=entity_tracker
+mod_name=Entity Tracker
 mod_license=GPLv3
 mod_version=1.21.1-0.0.1
-mod_group_id=nowebsite.makertechno.terra_ethud
+mod_group_id=nowebsite.makertechno.entity_tracker
 mod_authors=MakerTechno
 mod_description=Where did that entity just go? Did you see it?
