+++ src/main/java/nowebsite/makertechno/entity_tracker/algorithm/ProjectToAngleAndDistance.java
@@ -1,0 +1,8 @@
+package nowebsite.makertechno.entity_tracker.algorithm;
+
+import net.minecraft.world.phys.Vec2;
+import net.minecraft.world.phys.Vec3;
+
+public interface ProjectToAngleAndDistance {
+    Vec2 project(Vec3 vec);
+}
