+++ src/main/java/nowebsite/makertechno/entity_tracker/client/gui/TGuiLayers.java
@@ -1,0 +1,9 @@
+package nowebsite.makertechno.entity_tracker.client.gui;
+
+import net.minecraft.resources.ResourceLocation;
+import nowebsite.makertechno.entity_tracker.EntityTracker;
+
+public class TGuiLayers {
+
+    public static final ResourceLocation POINTER = ResourceLocation.fromNamespaceAndPath(EntityTracker.MOD_ID, "pointer");
+}
