+++ src/main/java/nowebsite/makertechno/entity_tracker/define/MappingStyle.java
@@ -1,0 +1,7 @@
+package nowebsite.makertechno.entity_tracker.define;
+
+public enum MappingStyle {
+    CENTER_RELATIVE,
+    HEAD_FLAT,
+    TRACK_FULL;
+}
