+++ src/main/java/nowebsite/makertechno/entity_tracker/define/TCursor.java
@@ -1,0 +1,3 @@
+package nowebsite.makertechno.entity_tracker.define;
+
+public record TCursor(TPointer pointer, TEntityIcon icon) {}
