--- src/main/java/nowebsite/makertechno/terra_ethud/algorithm/ProjectToAngleAndDistance.java
@@ -1,8 +1,0 @@
-package nowebsite.makertechno.terra_ethud.algorithm;
-
-import net.minecraft.world.phys.Vec2;
-import net.minecraft.world.phys.Vec3;
-
-public interface ProjectToAngleAndDistance {
-    Vec2 project(Vec3 vec);
-}
