--- src/main/java/nowebsite/makertechno/terra_ethud/client/event/TModClientGame.java
@@ -1,17 +1,0 @@
-package nowebsite.makertechno.terra_ethud.client.event;
-
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.client.event.ClientTickEvent;
-import nowebsite.makertechno.terra_ethud.TerraETHUD;
-import nowebsite.makertechno.terra_ethud.client.track.EntityTracker;
-import nowebsite.makertechno.terra_ethud.config.TConfig;
-
-@EventBusSubscriber(modid = TerraETHUD.MOD_ID, bus = EventBusSubscriber.Bus.GAME)
-public class TModClientGame {
-
-    @SubscribeEvent
-    public static void onClientTick(ClientTickEvent.Pre event) {
-        if (TConfig.available) EntityTracker.trigger();
-    }
-}
