--- src/main/java/nowebsite/makertechno/terra_ethud/client/gui/TGuiLayers.java
@@ -1,9 +1,0 @@
-package nowebsite.makertechno.terra_ethud.client.gui;
-
-import net.minecraft.resources.ResourceLocation;
-import nowebsite.makertechno.terra_ethud.TerraETHUD;
-
-public class TGuiLayers {
-
-    public static final ResourceLocation POINTER = ResourceLocation.fromNamespaceAndPath(TerraETHUD.MOD_ID, "pointer");
-}
