--- src/main/java/nowebsite/makertechno/terra_ethud/define/MappingStyle.java
@@ -1,7 +1,0 @@
-package nowebsite.makertechno.terra_ethud.define;
-
-public enum MappingStyle {
-    NORMAL,
-    HEAD_FLAT,
-    TRACK_FULL;
-}
