--- src/main/java/nowebsite/makertechno/terra_ethud/define/TCursor.java
@@ -1,3 +1,0 @@
-package nowebsite.makertechno.terra_ethud.define;
-
-public record TCursor(TPointer pointer, TEntityIcon icon) {}
