+++ src/main/resources/assets/entity_tracker/textures/gui/sprites/hud/ender_dragon.png
