+++ src/main/resources/assets/entity_tracker/textures/gui/sprites/hud/red.png
