+++ src/main/resources/assets/entity_tracker/textures/gui/sprites/hud/white.png
